agents/News TL;DR agent/requirements.txt
langgraph>=0.2.0
langchain-openai>=0.1.0
langchain-core>=0.2.0
pydantic>=2.0.0
python-dotenv>=1.0.0
newsapi-python>=0.2.7
beautifulsoup4>=4.12.0
requests>=2.31.0
lxml>=4.9.0
streamlit>=1.28.0


